requirements.txt
anyascii==0.3.3
asgiref==3.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
cryptography==43.0.3
defusedxml==0.7.1
Django==5.2.8
django-filter==25.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.7
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.7.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-allauth==65.1.0
dj-rest-auth==6.0.0
django-cors-headers==4.6.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
laces==0.1.2
modelsearch==1.1.1
openpyxl==3.1.5
pillow==12.0.0
pillow_heif==1.1.1
psycopg2==2.9.11
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wagtail==7.2
Willow==1.12.0
